moved deriv_display functionality into utils.
